chore(#120): Phase 254 — Build script DRY consolidation
Shared loadBrandingContext() extracted to branding.ts (DRY-M1),
VAULT_GUARD_PROMPT moved to hook-registry.ts (DRY-M3),
generateDomainOutputs() extracted to build-shared.ts (DRY-L1),
deprecated hookRegistry export removed (DRY-L3), dynamic imports
replaced with top-level (DX-L8), computeOutputCounts() for plugin
counting (DRY-L5).
