+++ .planning/STATE.md
@@ -66,4 +66,4 @@ _State updated: 2026-03-30 — v8.5.2 milestone complete (Phase 236 + Phase 237 
 
 ---
 
-_State generated from machine snapshot at 2026-03-31T16:51:09.263Z_
+_State generated from machine snapshot at 2026-03-31T17:00:26.064Z_
