+++ scripts/build-deploy.ts
@@ -43,8 +43,7 @@ import {
   buildErrorHandler,
 } from "./build-utils";
 import { resolvePackageRoot } from "../src/shared/__helpers/resolve-package-root";
-import { defaultBranding, validateBranding } from "./branding";
-import { sanitizeJsonParse } from "./sanitize";
+import { loadBrandingContext } from "./branding";
 
 // ---------------------------------------------------------------------------
 // Constants
@@ -58,80 +57,6 @@ const TEMPLATE_SOURCE_DIR = path.join(
   "harness",
   "claude",
 );
-
-// ---------------------------------------------------------------------------
-// Config loading
-// ---------------------------------------------------------------------------
-
-/**
- * Read branding config from `.planning/config.json` and compute derived values.
- *
- * Falls back to Luca defaults if the config file is missing or malformed.
- *
- * @returns Complete BrandingContext ready for template resolution
- */
-async function loadBrandingContext(): Promise<BrandingContext> {
-  const configPath = path.join(
-    resolvePackageRoot(),
-    ".planning",
-    "config.json",
-  );
-
-  let frameworkName = defaultBranding.frameworkName;
-  let commandPrefix = defaultBranding.commandPrefix;
-  let ticketPattern = defaultBranding.ticketPattern;
-  let placeholderTicket = defaultBranding.placeholderTicket;
-
-  try {
-    const configFile = Bun.file(configPath);
-    if (await configFile.exists()) {
-      const raw = sanitizeJsonParse(await configFile.text()) as Record<
-        string,
-        unknown
-      >;
-      const branding = (raw as Record<string, unknown>)?.branding as
-        | Record<string, string>
-        | undefined;
-      if (branding) {
-        frameworkName = branding.frameworkName ?? frameworkName;
-        commandPrefix = branding.commandPrefix ?? commandPrefix;
-        ticketPattern = branding.ticketPattern ?? ticketPattern;
-        placeholderTicket = branding.placeholderTicket ?? placeholderTicket;
-      }
-    }
-  } catch {
-    console.warn(
-      "Warning: Could not read .planning/config.json, using default branding",
-    );
-  }
-
-  // Validate resolved branding values (non-blocking: warn + continue on failure)
-  const validationResult = validateBranding({
-    frameworkName,
-    commandPrefix,
-    ticketPattern,
-    placeholderTicket,
-  });
-  if (!validationResult.valid) {
-    console.warn(
-      "Warning: Branding validation failed, continuing with current values:",
-    );
-    for (const [field, message] of Object.entries(validationResult.errors)) {
-      console.warn(`  ${field}: ${message}`);
-    }
-  }
-
-  return {
-    frameworkName,
-    commandPrefix,
-    commandSlash: `/${commandPrefix}`,
-    nameLowercase: frameworkName.toLowerCase(),
-    nameUppercase: frameworkName.toUpperCase(),
-    ticketPattern,
-    placeholderTicket,
-    ticketPatternJson: ticketPattern.replace(/\\/g, "\\\\"),
-  };
-}
 
 // ---------------------------------------------------------------------------
 // Main deploy function
@@ -163,7 +88,7 @@ export async function runDeploy(): Promise<{
   // =========================================================================
   // 1. Load branding context
   // =========================================================================
-  const branding = await loadBrandingContext();
+  const branding = await loadBrandingContext(packageRoot);
 
   // =========================================================================
   // 2. Resolve templates
