+++ scripts/check-drift.ts
@@ -21,42 +21,7 @@ import path from "path";
 import { readdirSync, statSync } from "fs";
 import { resolvePackageRoot } from "../src/shared/__helpers/resolve-package-root";
 import { transformOutputsToTemplates } from "../src/compilers";
-import { sanitizeJsonParse } from "./sanitize";
-import { defaultBranding } from "./branding";
-
-/**
- * Read branding config from .planning/config.json, falling back to defaults.
- * Mirrors the branding resolution logic in build-deploy.ts.
- */
-async function readBrandingContext(
-  projectDir: string,
-): Promise<Record<string, string>> {
-  let frameworkName = defaultBranding.frameworkName;
-  let commandPrefix = defaultBranding.commandPrefix;
-
-  try {
-    const configPath = path.join(projectDir, ".planning/config.json");
-    const text = await Bun.file(configPath).text();
-    const config = sanitizeJsonParse(text) as Record<string, unknown>;
-    const branding = config.branding as
-      | { frameworkName?: string; commandPrefix?: string }
-      | undefined;
-    if (branding) {
-      frameworkName = branding.frameworkName ?? frameworkName;
-      commandPrefix = branding.commandPrefix ?? commandPrefix;
-    }
-  } catch {
-    // Use defaults
-  }
-
-  return {
-    frameworkName,
-    commandPrefix,
-    commandSlash: `/${commandPrefix}`,
-    nameLowercase: frameworkName.toLowerCase(),
-    nameUppercase: frameworkName.toUpperCase(),
-  };
-}
+import { loadBrandingContext } from "./branding";
 
 /**
  * Resolve EJS template placeholders in content using branding context.
@@ -87,7 +52,14 @@ async function main() {
 
   // Apply branding transform to dist/claude/ entries (same as compile+deploy pipeline)
   // This converts `# lu` → `# /lu`, `lu-router` → `{prefix}-router`, etc.
-  const branding = await readBrandingContext(projectDir);
+  // loadBrandingContext returns a BrandingContext (named interface).
+  // Spread into Record<string, string> for dynamic key access in resolveBranding.
+  const brandingCtx = await loadBrandingContext(projectDir);
+  const branding: Record<string, string> = Object.fromEntries(
+    Object.entries(brandingCtx).filter(
+      (entry): entry is [string, string] => typeof entry[1] === "string",
+    ),
+  );
   const claudeEntries = new Map<string, string>();
   const nonClaudeEntries = new Map<string, string>();
